requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chromedriver-binary==133.0.6918.0.0
chromedriver-binary-auto==0.3.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.7.0
docopt==0.6.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
Flask==3.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaitaistruct==0.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.0.2
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydivert==2.1.0
Pygments==2.18.0
pymongo==4.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
selenium==4.27.1
selenium-wire==5.1.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
tabulate==0.9.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wsproto==1.2.0
yarg==0.1.9
zstandard==0.23.0
